fix: clear cart after successful order creation & mark order as paid on successful payment
