+++ routes/users.js
@@ -32,8 +32,8 @@ router.get('/membership/orders', auth, handleErrorAsync(usersController.getUserO
 router.get('/checkout', auth, handleErrorAsync(usersController.getCheckout));
 router.put('/checkout', auth, handleErrorAsync(usersController.putCheckout));
 router.post('/newebpay', auth, handleErrorAsync(usersController.postNeWebPay));
-router.get('/payment/callback', handleErrorAsync(usersController.getPaymentCallback));
-router.post('/payment/callback', handleErrorAsync(usersController.getPaymentCallback));
+// router.get('/payment/callback', handleErrorAsync(usersController.getPaymentCallback));
+router.post('/payment/callback', handleErrorAsync(usersController.postPaymentCallback));
 router.post('/payment/notify', handleErrorAsync(usersController.postPaymentNotify));
 router.post('/membership/discount', auth, handleErrorAsync(usersController.getDiscount)); // 試算優惠
 router.post('/membership/discount/usage', auth, handleErrorAsync(usersController.postDiscountUsage)); // 儲存使用紀錄
